Renames system control registers and associated tests
