+++ micromachine-test/CMakeLists.txt
@@ -132,11 +132,11 @@ declare_test(binary_helpers_tests.cpp)
 declare_test(exceptions/exception_prioritisation.cpp)
 declare_test(exceptions/exception_vector.cpp)
 declare_test(exceptions/exception_vector_reset.cpp)
-declare_test(systick_control_reg.cpp)
-declare_test(config_and_control_reg.cpp)
-declare_test(interrupt_and_reset_control_reg.cpp)
-declare_test(vtable_offset_reg.cpp)
-declare_test(icsr_reg.cpp)
+declare_test(registers/system_control/systick_control_reg_test.cpp)
+declare_test(registers/system_control/aircr_reg_test.cpp)
+declare_test(registers/system_control/icsr_reg_test.cpp)
+declare_test(registers/system_control/ccr_reg_test.cpp)
+declare_test(registers/system_control/vtor_reg_test.cpp)
 declare_test(nvic.cpp)
 declare_test(reset/reset_tests.cpp)
 declare_test(reset/exception_state_reset.cpp)
