+++ micromachine-test/registers/system_control/ccr_reg_test.cpp
@@ -3,8 +3,7 @@
 //
 
 
-#include "registers/system_control/config_and_control_reg.hpp"
-
+#include "registers/system_control/ccr_reg.hpp"
 
 #include <gtest/gtest.h>
