+++ micromachine/CMakeLists.txt
@@ -46,9 +46,9 @@ set(SRCS
 	registers/system_control/shpr2_reg.hpp
 	registers/system_control/shpr3_reg.hpp
 	registers/system_control/systick_control_reg.hpp
-	registers/system_control/config_and_control_reg.hpp
-	registers/system_control/interrupt_and_reset_control_reg.hpp
-	registers/system_control/vtable_offset_reg.hpp
+		registers/system_control/ccr_reg.hpp
+		registers/system_control/aircr_reg.hpp
+		registers/system_control/vtor_reg.hpp
 	registers/special_registers/apsr_reg.hpp
 	registers/special_registers/control_reg.hpp
 	registers/core_registers/core_registers.hpp
